refactored di package and changed module locations into their domains. Added another variable in models
