+++ .idea/misc.xml
@@ -39,7 +39,7 @@
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/question_item.xml" value="0.22" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/questions_screen.xml" value="0.131640625" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/quiz_congrats_dialog.xml" value="0.4466666666666667" />
-        <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/quiz_login_fragment.xml" value="0.22" />
+        <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/quiz_login_fragment.xml" value="0.16666666666666666" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/quiz_questions_fragment.xml" value="0.16666666666666666" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/start_screen.xml" value="0.17447916666666666" />
         <entry key="..\:/Users/Davit/Desktop/Quiz_App/app/src/main/res/layout/subject_item.xml" value="0.22" />
