+++ app/src/main/java/com/space/quiz_app/App.kt
@@ -1,10 +1,10 @@
 package com.space.quiz_app
 
 import android.app.Application
-import com.space.quiz_app.di.dataBaseModule
-import com.space.quiz_app.di.mapperModule
-import com.space.quiz_app.di.repositoryModule
-import com.space.quiz_app.di.viewModelModule
+import com.space.quiz_app.data.module.dataBaseModule
+import com.space.quiz_app.common.mapper.mapperModule
+import com.space.quiz_app.domain.module.repositoryModule
+import com.space.quiz_app.presentation.module.viewModelModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
