+++ app/src/main/java/com/space/quiz_app/common/mapper/Mapper.kt
@@ -1,4 +1,4 @@
-package com.space.quiz_app.common
+package com.space.quiz_app.common.mapper
 
 interface Mapper<in ModelA, out ModelB> {
     operator fun invoke(model: ModelA): ModelB
