+++ app/src/main/java/com/space/quiz_app/data/mapper/QuizUserEntityToDomainMapper.kt
@@ -1,6 +1,6 @@
 package com.space.quiz_app.data.mapper
 
-import com.space.quiz_app.common.Mapper
+import com.space.quiz_app.common.mapper.Mapper
 import com.space.quiz_app.data.local.QuizUserEntity
 import com.space.quiz_app.domain.model.QuizUserDomainModel
 
@@ -8,7 +8,8 @@ class QuizUserEntityToDomainMapper : Mapper<QuizUserEntity, QuizUserDomainModel>
     override fun invoke(model: QuizUserEntity): QuizUserDomainModel =
         with(model) {
             QuizUserDomainModel(
-                username = username
+                username = username,
+                gpa = gpa
             )
         }
 }
