+++ app/src/main/java/com/space/quiz_app/domain/model/QuizUserDomainModel.kt
@@ -1,6 +1,6 @@
 package com.space.quiz_app.domain.model
 
 data class QuizUserDomainModel(
-    val userId: Int? = null,
-    val username: String
+    val username: String,
+    val gpa: Float
 )
