+++ app/src/main/java/com/space/quiz_app/presentation/mapper/QuizUserUIToDomainMapper.kt
@@ -1,14 +1,15 @@
-package com.space.quiz_app.presentation.quiz_login_screen.mapper
+package com.space.quiz_app.presentation.mapper
 
-import com.space.quiz_app.common.Mapper
+import com.space.quiz_app.common.mapper.Mapper
 import com.space.quiz_app.domain.model.QuizUserDomainModel
-import com.space.quiz_app.presentation.quiz_login_screen.model.QuizUserUIModel
+import com.space.quiz_app.presentation.model.QuizUserUIModel
 
 class QuizUserUIToDomainMapper : Mapper<QuizUserUIModel, QuizUserDomainModel> {
     override fun invoke(model: QuizUserUIModel): QuizUserDomainModel =
         with(model) {
             QuizUserDomainModel(
-                username = username
+                username = username,
+                gpa = gpa
             )
         }
 }
