+++ app/src/main/java/com/space/quiz_app/presentation/model/QuizUserUIModel.kt
@@ -1,0 +1,6 @@
+package com.space.quiz_app.presentation.model
+
+data class QuizUserUIModel(
+    val username: String,
+    val gpa: Float
+)
