--- app/src/main/java/com/space/quiz_app/presentation/quiz_login_screen/model/QuizUserUIModel.kt
@@ -1,5 +1,0 @@
-package com.space.quiz_app.presentation.quiz_login_screen.model
-
-data class QuizUserUIModel(
-    val username: String
-)
