+++ app/src/main/res/layout/quiz_login_fragment.xml
@@ -6,7 +6,8 @@
     android:layout_height="match_parent"
     tools:context="com.space.quiz_app.presentation.quiz_login_screen.QuizLoginFragment">
 
-    <com.space.quiz_app.presentation.custom_view.QuizLoginBackgroundView
+    <com.space.quiz_app.presentation.custom_view.custom_view.QuizLoginBackgroundView
+        android:id="@+id/backgroundView"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintEnd_toEndOf="parent"
@@ -50,6 +51,7 @@
         android:paddingStart="@dimen/dimen_30_dp"
         android:paddingEnd="@dimen/dimen_30_dp"
         android:textAlignment="center"
+        android:background="@drawable/shp_message_edittext"
         app:layout_constraintBottom_toTopOf="@+id/startButton"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
